Working on making the nav toolbar work responsively.
So far I've made it so that when the screen reaches under 800px, the
class of the <ul> that contains the links will change.  Once that change
happens, then the <ul> with the links are moved over to a div and from
there, as of this moment, I'll switch the styling of the css.